Fix Raise Custom Exceptions
In file: scene_file_writer.py, method: add_audio_segment, an {Exception} or a {BaseException} was raised. This creates confusion since the exceptions are then differentiated only by the messages they carry. iCR suggested that a specific built-in exception or a custom exception is raised. For more information, read the Python guideline for raising exception.